requirements.txt
Django==1.11.7
djangorestframework==3.7.3
Markdown==2.6.9
pytz==2017.3
cgroups